+++ lib/iex/test/test_helper.exs
@@ -1,10 +1,11 @@
-:application.start(:iex)
+Application.start(:iex)
+Application.put_env(:iex, :colors, [enabled: false])
 ExUnit.start [trace: "--trace" in System.argv]
 
 defmodule IEx.Case do
+  use ExUnit.CaseTemplate
   @moduledoc false
 
-  #
   # Provides convenience functions for testing IEx-related functionality.
   # Use this module inside your test module like this:
   #
@@ -20,27 +21,25 @@ defmodule IEx.Case do
   # session, except colors are disabled by default and .iex files are not
   # loaded.
   #
-  # You can provide your own IEx.Options and a path to a .iex file as
+  # You can provide your own IEx configuration and a path to a .iex file as
   # additional arguments to the capture_iex function.
-  #
 
-  defmacro __using__(_) do
+  using do
     quote do
-      use ExUnit.Case, async: false
       import ExUnit.CaptureIO
       import unquote(__MODULE__)
-
-      setup do
-        opts = IEx.Options.get
-        IEx.Options.set :colors, [enabled: false]
-        {:ok, [iex_opts: opts]}
-      end
-
-      teardown context do
-        IEx.Options.set context[:iex_opts]
-        :ok
-      end
     end
+  end
+
+  setup do
+    opts = IEx.configuration |>
+           Keyword.take([:default_prompt, :alive_prompt, :inspect, :colors, :history_size])
+    {:ok, [iex_opts: opts]}
+  end
+
+  teardown context do
+    IEx.configure context[:iex_opts]
+    :ok
   end
 
   @doc """
@@ -54,9 +53,7 @@ defmodule IEx.Case do
   IEx.Server.start to be used in the normal .iex loading process.
   """
   def capture_iex(input, options \\ [], server_options \\ [], capture_prompt \\ false) do
-    Enum.each options, fn {opt, value} ->
-      IEx.Options.set(opt, value)
-    end
+    IEx.configure(options)
 
     ExUnit.CaptureIO.capture_io([input: input, capture_prompt: capture_prompt], fn ->
       server_options = Keyword.put_new(server_options, :dot_iex_path, "")